# src/main/java/edu/hei/school/restaurant/endpoint/rest/DishIngredientRequest.java
package edu.hei.school.restaurant.endpoint.rest;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class DishIngredientRequest {
    private Long ingredientId;
    private Double quantity;
} 