# src/main/java/edu/hei/school/restaurant/endpoint/rest/IngredientSummaryRest.java
package edu.hei.school.restaurant.endpoint.rest;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Builder
@Data
public class IngredientSummaryRest {
    private Long id;
    private String name;
    private Double currentPrice;
    private Double currentStock;
} 