# src/main/java/edu/hei/school/restaurant/endpoint/rest/OrderDishRequest.java
package edu.hei.school.restaurant.endpoint.rest;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class OrderDishRequest {
    private Long dishId;
    private Integer quantity;
} 