# src/main/java/edu/hei/school/restaurant/endpoint/rest/OrderDishRest.java
package edu.hei.school.restaurant.endpoint.rest;

public class OrderDishRest {

    public static Object builder() {
        // TODO Auto-generated method stub
        throw new UnsupportedOperationException("Unimplemented method 'builder'");
    }
} 